Split module graph stdlib gate tests
